requirements.txt
discord.py[voice]>=2.3.2
yt-dlp>=2024.1.1
python-dotenv>=1.0.0
PyNaCl>=1.5.0
aiohttp>=3.9.0
